Support loading AWS shared credentials
Update all classes that instantiate AWS objects to pass a region and
profile, then use a SharedCredentials object to open the connection.
This allows for passing in different AWS profiles as well as using
different regions without having to override via environment variables,
and conforms to current AWS best practices.